recommend

+++ administrators/views.py
@@ -6,7 +6,8 @@ from django.contrib.auth.models import Group, User
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic import View, DetailView
-from .forms import AddUserForm
+from .forms import AddManagerForm
+from django.contrib import messages 
 
 import os
 # from .forms import uploadMultiForm
@@ -30,9 +31,8 @@ def get_location_form(request):
     category_list = Category.objects.filter()
     return render(request, 'location/add.html',{"category_list": category_list})
 
-class add_location (LoginRequiredMixin, View):
-    login_url = '/login/'      
-    def post(self, request):   
+def add_location (request):
+    if request.user.is_authenticated:  
         if request.method =="POST":
             category = request.POST['category']
             name = request.POST['name']
@@ -66,31 +66,28 @@ class add_location (LoginRequiredMixin, View):
                                         timeend = timeend)
             location.save()
             return redirect('/administrators/location')
-            # if request.user.has_perm('Login.add_location'):
-            #     location.save()
-            #     return redirect('/administrators/location')
-            # else:
-            #     return render(request, 'common/error.html')
         else:
             return render(request, 'common/error.html')
+    else:
+        return redirect('/login')
+        
 
-class edit_location (LoginRequiredMixin, View):
-    login_url = '/login/' 
-    def post (request, id):
+def edit_location (request, id):
+    if request.user.is_authenticated: 
         loca = Location.objects.get(id=id)
         cate =  Category.objects.filter()
         if request.method =="POST":
-            # if len(request.FILES) !=0:
-            #     if len(loca.logo) > 0 and len(loca.image) > 0 :
-            #         os.remove(loca.logo.path)
-            #         os.remove(loca.image.path)  
-            #     if len(loca.logo) > 0 :
-            #         os.remove(loca.logo.path)
-            #     elif len(loca.image) > 0:
-            #         os.remove(loca.image.path)
+            if len(request.FILES) !=0:
+                if len(loca.logo) > 0 and len(loca.image) > 0 :
+                    os.remove(loca.logo.path)
+                    os.remove(loca.image.path)  
+                elif len(loca.logo) > 0 :
+                    os.remove(loca.logo.path)
+                elif len(loca.image) > 0:
+                    os.remove(loca.image.path)
                     
-            loca.logo = request.FILES['logo']
-            loca.image = request.FILES['image']
+                loca.logo = request.FILES['logo']
+                loca.image = request.FILES['image']
             loca.city = request.POST['city']
             loca.district = request.POST['district']
             loca.wardcommune = request.POST['ward']  
@@ -110,10 +107,11 @@ class edit_location (LoginRequiredMixin, View):
             messages.success(request, "Cập nhật thành công!")
             return redirect('/administrators/location')
         return render(request, 'location/edit.html',{"loca": loca, "cate": cate})
+    else:
+        return redirect('/login')
     
-class delete_location (LoginRequiredMixin, View):
-    login_url = '/login/' 
-    def post(request, id):
+def delete_location (request, id):
+    if request.user.is_authenticated:
         loca = Location.objects.get(id=id)
         if len(loca.logo) > 0 and len(loca.image) > 0:
             os.remove(loca.logo.path)
@@ -125,15 +123,16 @@ class delete_location (LoginRequiredMixin, View):
         loca.delete()
         messages.success(request, "Xoá thành công!")
         return redirect('/administrators/location')  
+    else:
+        return redirect('/login')
     
 # =============================IMAGES=========================
 def get_images_form(request):
     location_list = Location.objects.filter()
     return render(request, 'location/addImages.html',{"location_list": location_list})
 
-class add_images (LoginRequiredMixin, View):    
-    login_url = '/login/'     
-    def post (request):
+def add_images (request):    
+    if request.user.is_authenticated:
         if request.method =="POST":
             locationlis = request.POST['location_id']
             location = Location.objects.get(id=locationlis)
@@ -141,17 +140,24 @@ class add_images (LoginRequiredMixin, View):
             for img in images:
                 im = Images.objects.create(location_id=location,image = img)
                 im.save()
+            messages.success(request, "Thêm thành công!")
             return redirect('/administrators/formimages')
         else:
-            return render(request, 'common/error.html')    
+            messages.error(request, "Thêm mới thất bại!")
+            return render(request, 'common/error.html')  
+    else:
+        return redirect('/login')  
 
 def delete_images(request, images_id):
-    immge = Images.objects.get(id=images_id)
-    if len(immge.image) > 0:
-        os.remove(immge.image.path)
-    immge.delete()
-    messages.success(request, "Xoá thành công!")
-    return redirect('/administrators/formimages') 
+    if request.user.is_authenticated:
+        immge = Images.objects.get(id=images_id)
+        if len(immge.image) > 0:
+            os.remove(immge.image.path)
+        immge.delete()
+        messages.success(request, "Xoá thành công!")
+        return redirect('/administrators/formimages') 
+    else:
+        return redirect('/login')
       
 def viewImages(request):
     i = Images.objects.filter().order_by("-date")
@@ -170,28 +176,37 @@ def viewCategory(request):
     return render(request, 'category/index.html',{'page_obj': page_obj})
 
 def add_category(request):
-    if request.method =="POST":
-        name = request.POST['name']
-        cate = Category.objects.create(name = name)
-        cate.save()
-        return redirect('/administrators/category')
-    return render(request, 'category/add.html')
+    if request.user.is_authenticated:
+        if request.method =="POST":
+            name = request.POST['name']
+            cate = Category.objects.create(name = name)
+            cate.save()
+            messages.success(request, "Thêm mới thành công!")
+            return redirect('/administrators/category')
+        return render(request, 'category/add.html')
+    else: 
+        return redirect('/login')
 
 def edit_category (request, id):
-    category = Category.objects.get(id=id)
-    if request.method =="POST":
-        category.name = request.POST['name']
-        category.save()
-        messages.success(request, "Cập nhật thành công!")
-        return redirect('/administrators/category')
-    return render(request, 'category/edit.html',{"category": category})
+    if request.user.is_authenticated:
+        category = Category.objects.get(id=id)
+        if request.method =="POST":
+            category.name = request.POST['name']
+            category.save()
+            messages.success(request, "Cập nhật thành công!")
+            return redirect('/administrators/category')
+        return render(request, 'category/edit.html',{"category": category})
+    else:
+        return redirect('/login')
 
 def delete_category(request, id):
-    category = Category.objects.get(id=id)
-    category.delete()
-    messages.success(request, "Xoá thành công!")
-    return redirect('/administrators/category')  
-
+    if request.user.is_authenticated:
+        category = Category.objects.get(id=id)
+        category.delete()
+        messages.success(request, "Xoá thành công!")
+        return redirect('/administrators/category')  
+    else:
+        return redirect('/login')
 
 # ===================================================
 
@@ -205,8 +220,7 @@ def viewAccount(request):
     
 
 def viewUser(request,id ):
-    # group = Group.objects.get(id=id)
-    user = User.objects.filter(groups=id)
+    user = User.objects.filter(groups=id).order_by('-id')
     paginator = Paginator(user, 10)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
@@ -216,38 +230,17 @@ def get_group_list(request):
     group = Group.objects.all()
     return render(request, 'account/add.html',{'group': group})
 
-# class add_account(LoginRequiredMixin, View):
-#     login_url = '/login/'      
-#     def post(self, request):   
-#         if request.method =="POST":
-#             group = request.POST['group']
-#             username = request.POST['username']
-#             email = request.POST['email']
-#             password = request.POST['password']
-          
-#             group = Group.objects.get(id=group)
-#             account = User.objects.create(groups=group,
-#                                         username = username,
-#                                         email = email,
-#                                         password = password)
-#             account.save()
-#             return redirect('account/index.html')
-#         else:
-#             return render(request, 'common/error.html')
-
 def add_account(request):
-    form = AddUserForm()
-    if request.method == 'POST':
-        form = AddUserForm(request.POST)
-        if form.is_valid():
-            # form.save()
-            # user = form.save()
-            # group = Group.objects.get(name='Manager')
-            # group.user_set.add(user)
-            g = Group.objects.get(name='Manager')
-            users = form.save()
-            g.user_set.add(users)
-            return redirect('account_list')
-        else:
-            form = AddUserForm()
-    return render(request, 'account/add.html', {'form': form})
+    if request.user.is_authenticated:
+        form = AddManagerForm()
+        if request.method=='POST':
+            form = AddManagerForm(request.POST)
+            if form.is_valid():
+                form.save()
+                messages.success(request,'Thêm mới thành công')
+                return redirect('account_list')
+            else:
+                form = AddManagerForm()
+        return render(request,'account/add.html',{'form':form})
+    else:
+        return redirect('/login')
